A lot more buttons

+++ Assets/Scripts/CoinsStore.cs
@@ -8,18 +8,13 @@ using CodeControl;
 public class CoinsStore : MonoBehaviour
 {
 
-    int coinCoint = 0;
+    double coinCoint = 0;
     double count;
     void Awake()
     {
         Message.AddListener<CoinsUp>(OnCoinsUp);
         Message.AddListener<CoinsDown>(OnCoinsDown);
     }
-   //void OnDestroy()
-   //{
-   //    Message.RemoveListener<CoinsUp>(OnCoinsUp);
-   //     Message.RemoveListener<CoinsDown>(OnCoinsDown);
-   // }
    
     void TestCoins()
     {
@@ -36,7 +31,7 @@ public class CoinsStore : MonoBehaviour
     void OnCoinsDown(CoinsDown msg)
     {
         count = msg.count;
-        coinCoint = coinCoint - (int)count;
+        coinCoint = coinCoint - count;
         Message.Send(new CoinsUpdate(coinCoint));
     }
 }
